[#208] Put the tenant failure classifiers with the other exception matching
