+++ examples/university-multi-tenancy-examples/university-multi-tenancy-core/src/main/java/org/axonframework/examples/demo/multitenancy/shared/messaging/StatisticsSubscription.java
@@ -55,7 +55,7 @@ public final class StatisticsSubscription implements AutoCloseable {
 
     private StatisticsSubscription(QueryGateway queryGateway, TenantDescriptor tenant) {
         this.tenantId = tenant.tenantId();
-        this.subscription = Flux.from(Statistics.subscribeTo(queryGateway, tenant))
+        this.subscription = Flux.from(StatisticsQueries.subscribeTo(queryGateway, tenant))
                                 .doOnComplete(() -> {
                                     completed.set(true);
                                     logger.info("Tenant [{}]'s subscription completed: no seats left anywhere.",
